third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-get-with-virtual-authenticator.html

<!DOCTYPE html>
<title>Credential Manager: End-to-end tests for get() with a virtual authenticator.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/webauthn/authenticator.mojom.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/webauthn/virtual_authenticator.mojom.js"></script>
<script src="resources/test-inputs.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<body>
<script>

if (document.location.host != "subdomain.example.test:8443") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/credentialscontainer-get-with-virtual-authenticator.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

promise_test(async _ => {
  let authenticators = await navigator.credentials.test.authenticators();
  assert_equals(authenticators.length, 0);
  let testAuthenticator = await navigator.credentials.test.createAuthenticator();
  assert_true(await testAuthenticator.generateAndRegisterKey(ACCEPTABLE_CREDENTIAL_ID, "subdomain.example.test"));
}, "Set up the testing environment.");

promise_test(async t => {
  let testAuthenticator = await navigator.credentials.test.createAuthenticator();
  assert_true(await testAuthenticator.generateAndRegisterKey(ACCEPTABLE_CREDENTIAL_ID, "subdomain.example.test"));
  let keys = await testAuthenticator.registeredKeys();
  assert_equals(keys.length, 1);
  var customGetAssertionOptions = deepCopy(GET_CREDENTIAL_OPTIONS);
  var someOtherCredential = deepCopy(ACCEPTABLE_CREDENTIAL);
  someOtherCredential.id = new TextEncoder().encode("someOtherCredential");
  customGetAssertionOptions.allowCredentials = [someOtherCredential];

  return promise_rejects(t, "InvalidStateError",
    navigator.credentials.get({ publicKey : customGetAssertionOptions}));
}, "navigator.credentials.get() for unregistered device returns InvalidStateError");

promise_test(async t => {
  var customGetAssertionOptions = deepCopy(GET_CREDENTIAL_OPTIONS);
  var someOtherCredential = deepCopy(ACCEPTABLE_CREDENTIAL);
  someOtherCredential.id = new TextEncoder().encode("someOtherCredential");
  delete customGetAssertionOptions.allowCredentials;

  return promise_rejects(t, "NotSupportedError",
    navigator.credentials.get({ publicKey : customGetAssertionOptions}));
}, "navigator.credentials.get() with empty allowCredentials returns NotSupportedError");

promise_test(t => {
  return navigator.credentials.test.clearAuthenticators();
}, "Clean up testing environment.");

</script>
